+++ test/in-browser/spec/KefirSpecs.js
@@ -1682,16 +1682,47 @@ withOneSource('slidingWindow', {
   _init: function(args) {
     this._max = args[0];
     this._min = args[1] || 0;
-    this._cache = [];
+    this._buff = [];
   },
   _free: function() {
-    this._cache = null;
+    this._buff = null;
   },
   _handleValue: function(x, isCurrent) {
-    this._cache = slide(this._cache, x, this._max);
-    if (this._cache.length >= this._min) {
-      this._send(VALUE, this._cache, isCurrent);
+    this._buff = slide(this._buff, x, this._max);
+    if (this._buff.length >= this._min) {
+      this._send(VALUE, this._buff, isCurrent);
     }
+  }
+});
+
+
+
+
+// .bufferWhile([predicate])
+
+withOneSource('bufferWhile', {
+  _init: function(args) {
+    this._fn = args[0] || id;
+    this._buff = [];
+  },
+  _free: function() {
+    this._buff = null;
+  },
+  _flush: function(isCurrent) {
+    this._send(VALUE, this._buff, isCurrent);
+    this._buff = [];
+  },
+  _handleValue: function(x, isCurrent) {
+    this._buff.push(x);
+    if (!this._fn(x)) {
+      this._flush(isCurrent);
+    }
+  },
+  _handleEnd: function(x, isCurrent) {
+    if (this._buff.length !== 0) {
+      this._flush(isCurrent);
+    }
+    this._send(END, null, isCurrent);
   }
 });
 
@@ -21688,6 +21719,78 @@ module.exports = {
 };
 
 },{}],34:[function(require,module,exports){
+var Kefir, not3, prop, send, stream, _ref;
+
+_ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
+
+not3 = function(x) {
+  return x !== 3;
+};
+
+describe('bufferWhile', function() {
+  describe('stream', function() {
+    it('should return stream', function() {
+      return expect(stream().bufferWhile(not3)).toBeStream();
+    });
+    it('should activate/deactivate source', function() {
+      var a;
+      a = stream();
+      return expect(a.bufferWhile(not3)).toActivate(a);
+    });
+    it('should be ended if source was ended', function() {
+      return expect(send(stream(), ['<end>']).bufferWhile(not3)).toEmit(['<end:current>']);
+    });
+    return it('should work correctly', function() {
+      var a;
+      a = stream();
+      return expect(a.bufferWhile(not3)).toEmit([[3], [1, 2, 3], [4, 3], [3], [5, 6], '<end>'], function() {
+        return send(a, [3, 1, 2, 3, 4, 3, 3, 5, 6, '<end>']);
+      });
+    });
+  });
+  return describe('property', function() {
+    it('should return property', function() {
+      return expect(prop().bufferWhile(not3)).toBeProperty();
+    });
+    it('should activate/deactivate source', function() {
+      var a;
+      a = prop();
+      return expect(a.bufferWhile(not3)).toActivate(a);
+    });
+    it('should be ended if source was ended', function() {
+      expect(send(prop(), ['<end>']).bufferWhile(not3)).toEmit(['<end:current>']);
+      expect(send(prop(), [3, '<end>']).bufferWhile(not3)).toEmit([
+        {
+          current: [3]
+        }, '<end:current>'
+      ]);
+      return expect(send(prop(), [2, '<end>']).bufferWhile(not3)).toEmit([
+        {
+          current: [2]
+        }, '<end:current>'
+      ]);
+    });
+    return it('should work correctly', function() {
+      var a;
+      a = send(prop(), [3]);
+      expect(a.bufferWhile(not3)).toEmit([
+        {
+          current: [3]
+        }, [1, 2, 3], [4, 3], [3], [5, 6], '<end>'
+      ], function() {
+        return send(a, [1, 2, 3, 4, 3, 3, 5, 6, '<end>']);
+      });
+      a = send(prop(), [1]);
+      return expect(a.bufferWhile(not3)).toEmit([[1, 2, 3], [5, 6], '<end>'], function() {
+        return send(a, [2, 3, 5, 6, '<end>']);
+      });
+    });
+  });
+});
+
+
+
+},{"../test-helpers.coffee":89}],35:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -21823,7 +21926,7 @@ describe('bus', function() {
 
 
 
-},{"../test-helpers.coffee":88}],35:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],36:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -21858,7 +21961,7 @@ describe('changes', function() {
 
 
 
-},{"../test-helpers.coffee":88}],36:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],37:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref,
   __slice = [].slice;
 
@@ -21972,7 +22075,7 @@ describe('combine', function() {
 
 
 
-},{"../test-helpers.coffee":88}],37:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],38:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -22080,7 +22183,7 @@ describe('concat', function() {
 
 
 
-},{"../test-helpers.coffee":88}],38:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],39:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -22100,7 +22203,7 @@ describe('constant', function() {
 
 
 
-},{"kefir":90}],39:[function(require,module,exports){
+},{"kefir":91}],40:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22291,7 +22394,7 @@ describe('debounce', function() {
 
 
 
-},{"../test-helpers.coffee":88}],40:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],41:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22355,7 +22458,7 @@ describe('delay', function() {
 
 
 
-},{"../test-helpers.coffee":88}],41:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],42:[function(require,module,exports){
 var Kefir, minus, noop, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22455,7 +22558,7 @@ describe('diff', function() {
 
 
 
-},{"../test-helpers.coffee":88}],42:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],43:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -22481,7 +22584,7 @@ describe('emitter', function() {
 
 
 
-},{"kefir":90}],43:[function(require,module,exports){
+},{"kefir":91}],44:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22730,7 +22833,7 @@ describe('filterBy', function() {
 
 
 
-},{"../test-helpers.coffee":88}],44:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],45:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22817,7 +22920,7 @@ describe('filter', function() {
 
 
 
-},{"../test-helpers.coffee":88}],45:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],46:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -22952,7 +23055,7 @@ describe('flatMapConcat', function() {
 
 
 
-},{"../test-helpers.coffee":88}],46:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],47:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23107,7 +23210,7 @@ describe('flatMapFirst', function() {
 
 
 
-},{"../test-helpers.coffee":88}],47:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],48:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23241,7 +23344,7 @@ describe('flatMapLatest', function() {
 
 
 
-},{"../test-helpers.coffee":88}],48:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],49:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23372,7 +23475,7 @@ describe('flatMapConcurLimit', function() {
 
 
 
-},{"../test-helpers.coffee":88}],49:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],50:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23535,7 +23638,7 @@ describe('flatMap', function() {
 
 
 
-},{"../test-helpers.coffee":88}],50:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],51:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -23653,7 +23756,7 @@ describe('flatten', function() {
 
 
 
-},{"../test-helpers.coffee":88}],51:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],52:[function(require,module,exports){
 var Kefir, activate, deactivate, _ref;
 
 _ref = require('../test-helpers.coffee'), activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23735,7 +23838,7 @@ describe('fromBinder', function() {
 
 
 
-},{"../test-helpers.coffee":88}],52:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],53:[function(require,module,exports){
 var Kefir, activate, deactivate, _ref;
 
 _ref = require('../test-helpers.coffee'), activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23798,7 +23901,7 @@ describe('fromCallback', function() {
 
 
 
-},{"../test-helpers.coffee":88}],53:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],54:[function(require,module,exports){
 var Kefir, activate, deactivate, _ref;
 
 _ref = require('../test-helpers.coffee'), activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23931,7 +24034,7 @@ describe('fromEvent', function() {
 
 
 
-},{"../test-helpers.coffee":88}],54:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],55:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -23951,7 +24054,7 @@ describe('fromPoll', function() {
 
 
 
-},{"kefir":90}],55:[function(require,module,exports){
+},{"kefir":91}],56:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -23967,7 +24070,7 @@ describe('interval', function() {
 
 
 
-},{"kefir":90}],56:[function(require,module,exports){
+},{"kefir":91}],57:[function(require,module,exports){
 var $, Kefir, countListentrs, inBrowser, withDOM, _ref;
 
 _ref = require('../test-helpers.coffee'), withDOM = _ref.withDOM, inBrowser = _ref.inBrowser, Kefir = _ref.Kefir;
@@ -24277,7 +24380,7 @@ if (!inBrowser) {
 
 
 
-},{"../test-helpers.coffee":88,"addons/kefir-jquery":89,"jquery":6}],57:[function(require,module,exports){
+},{"../test-helpers.coffee":89,"addons/kefir-jquery":90,"jquery":6}],58:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -24293,7 +24396,7 @@ describe('later', function() {
 
 
 
-},{"kefir":90}],58:[function(require,module,exports){
+},{"kefir":91}],59:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -24370,7 +24473,7 @@ describe('mapEnd', function() {
 
 
 
-},{"../test-helpers.coffee":88}],59:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],60:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -24428,7 +24531,7 @@ describe('map', function() {
 
 
 
-},{"../test-helpers.coffee":88}],60:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],61:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -24524,7 +24627,7 @@ describe('merge', function() {
 
 
 
-},{"../test-helpers.coffee":88}],61:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],62:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -24540,7 +24643,7 @@ describe('never', function() {
 
 
 
-},{"kefir":90}],62:[function(require,module,exports){
+},{"kefir":91}],63:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -24639,7 +24742,7 @@ describe('pool', function() {
 
 
 
-},{"../test-helpers.coffee":88}],63:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],64:[function(require,module,exports){
 var Kefir, activate, prop, send, _ref;
 
 _ref = require('../test-helpers.coffee'), prop = _ref.prop, send = _ref.send, activate = _ref.activate, Kefir = _ref.Kefir;
@@ -24799,7 +24902,7 @@ describe('Property', function() {
 
 
 
-},{"../test-helpers.coffee":88}],64:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],65:[function(require,module,exports){
 var Kefir, minus, noop, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -24895,7 +24998,7 @@ describe('reduce', function() {
 
 
 
-},{"../test-helpers.coffee":88}],65:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],66:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -24914,7 +25017,7 @@ describe('repeatedly', function() {
 
 
 
-},{"kefir":90}],66:[function(require,module,exports){
+},{"kefir":91}],67:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref,
   __slice = [].slice;
 
@@ -25051,7 +25154,7 @@ describe('sampledBy', function() {
 
 
 
-},{"../test-helpers.coffee":88}],67:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],68:[function(require,module,exports){
 var Kefir, minus, noop, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25141,7 +25244,7 @@ describe('scan', function() {
 
 
 
-},{"../test-helpers.coffee":88}],68:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],69:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -25160,7 +25263,7 @@ describe('sequentially', function() {
 
 
 
-},{"kefir":90}],69:[function(require,module,exports){
+},{"kefir":91}],70:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25236,7 +25339,7 @@ describe('skipDuplicates', function() {
 
 
 
-},{"../test-helpers.coffee":88}],70:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],71:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25295,7 +25398,7 @@ describe('mapEnd', function() {
 
 
 
-},{"../test-helpers.coffee":88}],71:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],72:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir, activate = _ref.activate, deactivate = _ref.deactivate;
@@ -25549,7 +25652,7 @@ describe('skipUntilBy', function() {
 
 
 
-},{"../test-helpers.coffee":88}],72:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],73:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir, deactivate = _ref.deactivate, activate = _ref.activate;
@@ -25806,7 +25909,7 @@ describe('skipWhileBy', function() {
 
 
 
-},{"../test-helpers.coffee":88}],73:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],74:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25934,7 +26037,7 @@ describe('skipWhile', function() {
 
 
 
-},{"../test-helpers.coffee":88}],74:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],75:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26034,7 +26137,7 @@ describe('skip', function() {
 
 
 
-},{"../test-helpers.coffee":88}],75:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],76:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26130,7 +26233,7 @@ describe('slidingWindow', function() {
 
 
 
-},{"../test-helpers.coffee":88}],76:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],77:[function(require,module,exports){
 var Kefir, activate, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, send = _ref.send, activate = _ref.activate, Kefir = _ref.Kefir;
@@ -26321,7 +26424,7 @@ describe('Stream', function() {
 
 
 
-},{"../test-helpers.coffee":88}],77:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],78:[function(require,module,exports){
 var Kefir, expectToBehaveAsMap, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26573,7 +26676,7 @@ describe('awaiting', function() {
 
 
 
-},{"../test-helpers.coffee":88}],78:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],79:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26817,7 +26920,7 @@ describe('takeUntilBy', function() {
 
 
 
-},{"../test-helpers.coffee":88}],79:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],80:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27066,7 +27169,7 @@ describe('takeWhileBy', function() {
 
 
 
-},{"../test-helpers.coffee":88}],80:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],81:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27203,7 +27306,7 @@ describe('takeWhile', function() {
 
 
 
-},{"../test-helpers.coffee":88}],81:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],82:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27288,7 +27391,7 @@ describe('take', function() {
 
 
 
-},{"../test-helpers.coffee":88}],82:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],83:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27562,7 +27665,7 @@ describe('throttle', function() {
 
 
 
-},{"../test-helpers.coffee":88}],83:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],84:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27651,7 +27754,7 @@ describe('timestamp', function() {
 
 
 
-},{"../test-helpers.coffee":88}],84:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],85:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27697,7 +27800,7 @@ describe('toProperty', function() {
 
 
 
-},{"../test-helpers.coffee":88}],85:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],86:[function(require,module,exports){
 var Kefir, comp, noop, prop, send, stream, testWithLib, _ref,
   __slice = [].slice;
 
@@ -27947,7 +28050,7 @@ describe('transduce', function() {
 
 
 
-},{"../test-helpers.coffee":88,"transducers-js":32,"transducers.js":33}],86:[function(require,module,exports){
+},{"../test-helpers.coffee":89,"transducers-js":32,"transducers.js":33}],87:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -28064,7 +28167,7 @@ describe('withHandler', function() {
 
 
 
-},{"../test-helpers.coffee":88}],87:[function(require,module,exports){
+},{"../test-helpers.coffee":89}],88:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -28090,7 +28193,7 @@ describe('withInterval', function() {
 
 
 
-},{"kefir":90}],88:[function(require,module,exports){
+},{"kefir":91}],89:[function(require,module,exports){
 var Kefir, getCurrent, logItem, sinon, _activateHelper,
   __slice = [].slice;
 
@@ -28339,7 +28442,7 @@ beforeEach(function() {
 
 
 
-},{"../dist/kefir":1,"sinon":7}],89:[function(require,module,exports){
+},{"../dist/kefir":1,"sinon":7}],90:[function(require,module,exports){
 /*! An addon for Kefir.js v0.4.0
  *  https://github.com/pozadi/kefir
  */
@@ -28391,6 +28494,6 @@ beforeEach(function() {
 
 }(this));
 
-},{"jquery":6,"kefir":90}],90:[function(require,module,exports){
+},{"jquery":6,"kefir":91}],91:[function(require,module,exports){
 module.exports=require(1)
-},{"/Users/anon/projects/my/kefir/dist/kefir.js":1}]},{},[34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87]);
+},{"/Users/anon/projects/my/kefir/dist/kefir.js":1}]},{},[34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88]);